Created a Review table/object, and attempted to connect to a trail
I created a Review model, controller, and repository. I tried to make it so that a Trail object contained a list of reviews, but I couldn't figure out how to add a trail id to review. So we can post and get review objects but they won't contain a trail id yet. Same thing for the hash set of reviews.